split lib in few files
