+++ README.md
@@ -9,7 +9,7 @@ rspec-html-matchers
 Goals
 -----
 
-* for testing **complicated** html output, for simple matching consider use:
+* designed for testing **complex** html output. If you plan to perform simple matching, consider using:
   * [assert_select](http://api.rubyonrails.org/classes/ActionDispatch/Assertions/SelectorAssertions.html#method-i-assert_select)
   * [matchers provided out of the box in rspec-rails](https://www.relishapp.com/rspec/rspec-rails/v/2-11/docs/view-specs/view-spec)
   * [matchers provided by capybara](http://rdoc.info/github/jnicklas/capybara/Capybara/Node/Matchers)
