+++ lib/rspec-html-matchers/version.rb
@@ -1,0 +1,5 @@
+module RSpecHtmlMatchers
+  module Version
+    VERSION = '0.9.1'
+  end
+end
